Refactored lifecycle package
